requirements.txt
requests
python-socketio[client]
pyserial
aiortc
opencv-python
asyncio
av
aiohttp
pyaudio
openai
python-dotenv
elevenlabs
ultralytics
smbus2
pandas
torch
transformers
huggingface-hub
accelerate